requirements.txt
Django==6.0.0
djangorestframework==3.14.0
django-rest-framework-simplejwt==5.3.2
django-cors-headers==4.3.1
Pillow==10.1.0
qrcode==7.4.2
twilio==8.10.0
psycopg2-binary==2.9.9
python-decouple==3.8
